Add shell script to automatically build angular project with dev and prod modes separated, fix environment variables
