+++ src/environments/environment.ts
@@ -3,16 +3,9 @@ import { appVersion } from './version';
 export const environment = {
   production: false,
   appVersion: appVersion,
-  //baseUrl: 'https://api.ecrin-rms.org',
+  appTitle: 'crDSR: Data Sharing Repository',
   baseUrl: 'https://api-dsrdev.ecrin.org/api',
-  // baseUrl: 'https://api-test.ecrin-rms.org/api',
-  // baseUrl: 'http://localhost:7220/api',
-  wsBaseUrl: 'wss://api-test.ecrin-rms.org/',
-
-  USERDATA_KEY: 'authf649fc9a5f55',
-  isMockEnabled: true,
-
-  hostname: 'http://localhost:5001/',
+  wsBaseUrl: 'wss://api-dsrdev.ecrin.org/',  // Not up
   apiUrl: 'api',
   apiVersion: 'v1'
 };
