+++ update_build.sh
@@ -1,0 +1,83 @@
+#!/bin/bash
+
+# Source of some of the code: https://medium.com/@wujido20/handling-flags-in-bash-scripts-4b06b4d0ed04
+
+# Default variable values
+production=false
+verbose=false
+build_path="dist/rms-portal"
+backup_path="dist/rms-portal-backup"
+target_path="/var/www/rms-portal-new/dist/rms-portal"
+
+
+usage() {
+    echo "Usage: $0 [OPTIONS]"
+    echo "Options:"
+    echo " -h, --help                   Display this help message"
+    echo " -v, --verbose                Enable verbose mode (outputs commands)"
+    echo " -p, --production             Build in production mode instead of default dev mode"
+    echo " -b, --build_path [path]      Path of angular build result (probably shouldn't change), default: $build_path"
+    echo " -s, --backup_path [path]     Path of previous build to be backed up, default: $backup_path"
+    echo " -t, --target_path [path]     Path of directory listed in the Nginx config file to move the build to, default: $target_path"
+}
+
+build() {
+    if [[ "$verbose" = true ]]; then
+        set -x
+    fi
+    if [[ "$production" = true ]]; then
+        npm run build_prod
+    else
+        npm run build_dev
+    fi
+    sudo rm -rf $backup_path
+    if [[ -d $target_path ]]; then
+        sudo mkdir -p $backup_path
+        sudo mv $target_path $backup_path
+    else
+        sudo mkdir -p $target_path
+    fi
+    if [[ -d $build_path ]]; then
+        sudo mv $build_path $target_path
+    else
+        echo "Error: build directory ($build_path) does not exist."
+        exit 1
+    fi
+    systemctl restart nginx
+}
+
+# Parsing command-line arguments
+while [ $# -gt 0 ]; do
+    case $1 in
+        -h | --help)
+            usage
+            exit 0
+            ;;
+        -v | --verbose)
+            verbose=true
+            ;;
+        -p | --production)
+            production=true
+            ;;
+        -b | --build_path)
+            build_path=$@
+            shift
+            ;;
+        -s | --backup_path)
+            backup_path=$@
+            shift
+            ;;
+        -t | --target_path)
+            target_path=$@
+            shift
+            ;;
+        *)
+            echo "Invalid option: $1" >&2
+            usage
+            exit 1
+            ;;
+    esac
+    shift
+done
+
+build
